refactor(guidance): extract teleport detection into GuidanceMovementTracker
Per project rule "many small files > few large files, 200-400 lines
typical, 800 max": the @Subscribe onGameTick added in the previous
commit pushed GuidanceEventRouter from 553 -> 649 lines and to 11
distinct event responsibilities. New contributors would have to read
the whole router to find any one handler.

Extract the per-tick player-movement tracking into a small focused
class GuidanceMovementTracker (~160 lines, single @Subscribe, no
setters, constructor injection only). Router returns to its
pre-#381 size of 553 lines.

This also creates a natural home for #378 (bidirectional step
regression on player leaving step location), which is the same
trigger class -- player movement during active guidance -- and would
otherwise force another inflation of the router.

Wiring: plugin @Injects the tracker and eventBus.register/unregister
alongside the existing guidanceEventRouter pair. No public API change.

Tests migrated 1:1 from GuidanceEventRouterTest to a new
GuidanceMovementTrackerTest in the matching package; no test logic
changed, only the SUT field and class location. 555/555 still pass.
