+++ docs/in-game-validation-log.md
@@ -122,7 +122,7 @@ A proper architectural fix (pre-resolved `List<RequiredItemDisplay>` passed from
 
 Closes cha-ndler/collection-log-helper#381. After teleporting away mid-guidance (e.g. Mort'ton → Grand Exchange via bank teleport, then back), the in-game hint arrow stayed cleared at the new step's target location. `client.setHintArrow(WorldPoint)` is bound to the loaded scene — when the scene reloaded, the arrow target was no longer rendered and walking back into range did not auto-restore it. Before this fix, only step transitions and the "Show Hint Arrow" config toggle re-set the arrow; neither fires on a teleport.
 
-Fix: `GuidanceEventRouter` now subscribes to `GameTick` and tracks the player's WorldPoint each tick while guidance is active. When the single-tick Chebyshev delta exceeds 10 tiles (running caps at 2 tiles/tick — any single-tick jump >10 is unambiguously a teleport) or the plane changes, it calls `GuidanceOverlayCoordinator.refreshHintArrow()`. That existing method already gates on `config.showHintArrow()` and the active step's world target, so the router fires unconditionally on teleport and lets the coordinator decide whether to set, clear, or no-op.
+Fix: introduces a new small focused class `com.collectionloghelper.guidance.GuidanceMovementTracker` (~160 lines, single responsibility) that subscribes to `GameTick` and tracks the player's WorldPoint each tick while guidance is active. When the single-tick Chebyshev delta exceeds 10 tiles (running caps at 2 tiles/tick — any single-tick jump >10 is unambiguously a teleport) or the plane changes, it calls `GuidanceOverlayCoordinator.refreshHintArrow()`. That existing method already gates on `config.showHintArrow()` and the active step's world target, so the tracker fires unconditionally on teleport and lets the coordinator decide whether to set, clear, or no-op. Extracted as a separate class (rather than another handler on `GuidanceEventRouter`, which already has 10 subscribers and 549 lines) so future movement-driven guidance behavior (e.g. step regression on player-left-step-location for cha-ndler/collection-log-helper#378) can live alongside it.
 
 | # | Test | Status | Notes |
 |---|---|---|---|
