+++ src/test/java/com/collectionloghelper/lifecycle/GuidanceEventRouterTest.java
@@ -47,10 +47,8 @@ import net.runelite.api.MenuAction;
 import net.runelite.api.MenuEntry;
 import net.runelite.api.NPC;
 import net.runelite.api.Player;
-import net.runelite.api.coords.WorldPoint;
 import net.runelite.api.events.ActorDeath;
 import net.runelite.api.events.AnimationChanged;
-import net.runelite.api.events.GameTick;
 import net.runelite.api.events.HitsplatApplied;
 import net.runelite.api.events.InteractingChanged;
 import net.runelite.api.events.MenuEntryAdded;
@@ -957,183 +955,5 @@ public class GuidanceEventRouterTest
 
 		// Must not throw
 		router.onConfigChanged(event);
-	}
-
-	// ========================================================================
-	// onGameTick — hint-arrow refresh after long-distance teleport (#381)
-	// ========================================================================
-	//
-	// Regression contract for cha-ndler/collection-log-helper#381:
-	// the router must detect single-tick position deltas above
-	// TELEPORT_DISTANCE_THRESHOLD (or plane changes) while guidance is active
-	// and call GuidanceOverlayCoordinator#refreshHintArrow(). It must NOT
-	// fire on ordinary walking, the first tick after activation, or when
-	// guidance is inactive — those would either churn or break the existing
-	// activation-time skip-chain.
-
-	/** First tick after activation establishes the baseline; no refresh fires. */
-	@Test
-	public void onGameTick_firstTickAfterActivation_doesNotRefresh()
-	{
-		when(guidanceSequencer.isActive()).thenReturn(true);
-		when(localPlayer.getWorldLocation()).thenReturn(new WorldPoint(3200, 3200, 0));
-
-		router.onGameTick(new GameTick());
-
-		verify(guidanceCoordinator, never()).refreshHintArrow();
-	}
-
-	/** Walking one tile between ticks must not trigger a refresh. */
-	@Test
-	public void onGameTick_walkingOneTile_doesNotRefresh()
-	{
-		when(guidanceSequencer.isActive()).thenReturn(true);
-		when(localPlayer.getWorldLocation())
-			.thenReturn(new WorldPoint(3200, 3200, 0))
-			.thenReturn(new WorldPoint(3201, 3200, 0));
-
-		router.onGameTick(new GameTick());
-		router.onGameTick(new GameTick());
-
-		verify(guidanceCoordinator, never()).refreshHintArrow();
-	}
-
-	/** Running two tiles between ticks (max walk speed) must not trigger a refresh. */
-	@Test
-	public void onGameTick_runningTwoTiles_doesNotRefresh()
-	{
-		when(guidanceSequencer.isActive()).thenReturn(true);
-		when(localPlayer.getWorldLocation())
-			.thenReturn(new WorldPoint(3200, 3200, 0))
-			.thenReturn(new WorldPoint(3202, 3202, 0));
-
-		router.onGameTick(new GameTick());
-		router.onGameTick(new GameTick());
-
-		verify(guidanceCoordinator, never()).refreshHintArrow();
-	}
-
-	/**
-	 * Long-distance teleport (Mort'ton → GE is ~80 tiles) must trigger refresh.
-	 * This is the #381 regression case.
-	 */
-	@Test
-	public void onGameTick_longDistanceTeleport_refreshesHintArrow()
-	{
-		when(guidanceSequencer.isActive()).thenReturn(true);
-		when(localPlayer.getWorldLocation())
-			.thenReturn(new WorldPoint(3492, 3299, 0))   // Mort'ton
-			.thenReturn(new WorldPoint(3164, 3486, 0));  // Grand Exchange
-
-		router.onGameTick(new GameTick());
-		router.onGameTick(new GameTick());
-
-		verify(guidanceCoordinator).refreshHintArrow();
-	}
-
-	/** Plane change (climbing stairs, dungeon entry) must trigger refresh. */
-	@Test
-	public void onGameTick_planeChange_refreshesHintArrow()
-	{
-		when(guidanceSequencer.isActive()).thenReturn(true);
-		when(localPlayer.getWorldLocation())
-			.thenReturn(new WorldPoint(3200, 3200, 0))
-			.thenReturn(new WorldPoint(3200, 3200, 1));
-
-		router.onGameTick(new GameTick());
-		router.onGameTick(new GameTick());
-
-		verify(guidanceCoordinator).refreshHintArrow();
-	}
-
-	/** Exactly at threshold (10 tiles) is still treated as walking; no refresh. */
-	@Test
-	public void onGameTick_deltaAtThreshold_doesNotRefresh()
-	{
-		when(guidanceSequencer.isActive()).thenReturn(true);
-		when(localPlayer.getWorldLocation())
-			.thenReturn(new WorldPoint(3200, 3200, 0))
-			.thenReturn(new WorldPoint(3210, 3200, 0));
-
-		router.onGameTick(new GameTick());
-		router.onGameTick(new GameTick());
-
-		verify(guidanceCoordinator, never()).refreshHintArrow();
-	}
-
-	/** One tile above threshold fires refresh — pins the boundary. */
-	@Test
-	public void onGameTick_deltaJustOverThreshold_refreshes()
-	{
-		when(guidanceSequencer.isActive()).thenReturn(true);
-		when(localPlayer.getWorldLocation())
-			.thenReturn(new WorldPoint(3200, 3200, 0))
-			.thenReturn(new WorldPoint(3211, 3200, 0));
-
-		router.onGameTick(new GameTick());
-		router.onGameTick(new GameTick());
-
-		verify(guidanceCoordinator).refreshHintArrow();
-	}
-
-	/** Inactive guidance: no refresh fires regardless of position delta. */
-	@Test
-	public void onGameTick_guidanceInactive_doesNotRefresh()
-	{
-		when(guidanceSequencer.isActive()).thenReturn(false);
-
-		router.onGameTick(new GameTick());
-		router.onGameTick(new GameTick());
-
-		verify(guidanceCoordinator, never()).refreshHintArrow();
-		// Player location must not even be queried when inactive
-		verify(localPlayer, never()).getWorldLocation();
-	}
-
-	/**
-	 * Guidance deactivates mid-tracking — when it reactivates and a teleport
-	 * happens on the very first active tick, no false refresh fires (baseline
-	 * was reset on the inactive tick).
-	 */
-	@Test
-	public void onGameTick_deactivationResetsBaseline_noFalseRefreshOnResume()
-	{
-		when(guidanceSequencer.isActive())
-			.thenReturn(true)   // tick 1: active, sets baseline at Mort'ton
-			.thenReturn(false)  // tick 2: inactive, clears baseline
-			.thenReturn(true);  // tick 3: re-active at GE, must NOT diff vs Mort'ton
-		when(localPlayer.getWorldLocation())
-			.thenReturn(new WorldPoint(3492, 3299, 0))
-			.thenReturn(new WorldPoint(3164, 3486, 0));
-
-		router.onGameTick(new GameTick());  // baseline = Mort'ton
-		router.onGameTick(new GameTick());  // inactive, baseline cleared
-		router.onGameTick(new GameTick());  // re-active, baseline re-established at GE
-
-		verify(guidanceCoordinator, never()).refreshHintArrow();
-	}
-
-	/** Null player (e.g. mid-login frame) is a no-op, no NPE. */
-	@Test
-	public void onGameTick_nullLocalPlayer_doesNotThrow()
-	{
-		when(guidanceSequencer.isActive()).thenReturn(true);
-		when(client.getLocalPlayer()).thenReturn(null);
-
-		router.onGameTick(new GameTick());
-
-		verify(guidanceCoordinator, never()).refreshHintArrow();
-	}
-
-	/** Null world location (transient client state) is a no-op, no NPE. */
-	@Test
-	public void onGameTick_nullWorldLocation_doesNotThrow()
-	{
-		when(guidanceSequencer.isActive()).thenReturn(true);
-		when(localPlayer.getWorldLocation()).thenReturn(null);
-
-		router.onGameTick(new GameTick());
-
-		verify(guidanceCoordinator, never()).refreshHintArrow();
 	}
 }
